rename package, improviments in jpa
