+++ server/src/main/java/br/com/danilopaixao/ws/core/BaseEntity.java
@@ -1,6 +1,5 @@
-package br.com.damilopaixao.ws.core;
+/**package br.com.danilopaixao.ws.core;
 
-/***
 import java.io.Serializable;
 import java.time.LocalDate;
