+++ server/src/main/java/br/com/danilopaixao/ws/core/LocalDateTimeConverter.java
@@ -1,4 +1,4 @@
-package br.com.damilopaixao.ws.core;
+package br.com.danilopaixao.ws.core;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
