+++ server/src/main/java/br/com/danilopaixao/ws/core/MoreInfo.java
@@ -1,4 +1,4 @@
-package br.com.damilopaixao.ws.core;
+package br.com.danilopaixao.ws.core;
 
 import java.lang.management.ManagementFactory;
 import java.lang.management.RuntimeMXBean;
