+++ server/src/main/java/br/com/danilopaixao/ws/core/SupplierExceptions.java
@@ -1,4 +1,4 @@
-package br.com.damilopaixao.ws.core;
+package br.com.danilopaixao.ws.core;
 
 import java.io.Serializable;
 import java.util.function.Supplier;
